requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
numpy>=1.26.0
opencv-python>=4.10.0
insightface>=0.7.3
onnxruntime>=1.16.0
scipy>=1.12.0
Pillow>=10.2.0
setuptools>=70.0.0
wheel>=0.43.0
stripe>=5.4.0
python-dotenv>=1.0.0